requirements.txt
# Core dependencies
pygame-ce>=2.5.0
pygame-gui>=0.6.0

# Game utilities  
pillow>=10.0.0

# Development (optional)
pytest>=7.0.0
black>=23.0.0
